Feat: get 함수 service 폴더로 복귀
